move translations filterting to AppTranslationGenerator
